Added python interface for image registration - currently has basic and single layer registration. Tested using Python 3.3.2. Added new test datasets for image registration, single band version of CASI image and offset by 3 pixels. TODO: Add warp commands

+++ src/cmds/RSGISCmdImageRegistration.cpp
@@ -124,8 +124,8 @@ namespace rsgis{ namespace cmds {
     
     void excecuteSingleLayerConnectedRegistration(std::string inputReferenceImage, std::string inputFloatingmage, int gcpGap,
                                                   float metricThreshold, int windowSize, int searchArea, float stdDevRefThreshold,
-                                                  float stdDevFloatThreshold, int subPixelResolution, float distanceThreshold,
-                                                  int maxNumIterations, float moveChangeThreshold, int pSmoothness, unsigned int metricTypeInt,
+                                                  float stdDevFloatThreshold, int subPixelResolution, int distanceThreshold,
+                                                  int maxNumIterations, float moveChangeThreshold, float pSmoothness, unsigned int metricTypeInt,
                                                   unsigned int outputType, std::string outputGCPFile)
     {
